Rename help component submodule.
