+++ client/src/component/Help/ManualContainer.tsx
